LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/iframe.meta/origin-when-crossorigin/iframe-tag.http-expected.txt

CONSOLE MESSAGE: Failed to set referrer policy: The value 'origin-when-crossorigin' is not one of 'no-referrer', 'no-referrer-when-downgrade', 'same-origin', 'origin', 'strict-origin', 'origin-when-cross-origin', 'strict-origin-when-cross-origin' or 'unsafe-url'.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/document.py?redirection=keep-origin&action=purge&key=bc253e80-7849-4d60-ba43-244af8fed88d&path=%2Fmixed-content


Harness Error (TIMEOUT), message = null

TIMEOUT Referrer Policy: Expects origin for iframe-tag to cross-http origin and keep-origin redirection from http context. Test timed out
NOTRUN Referrer Policy: Expects origin for iframe-tag to cross-http origin and no-redirect redirection from http context.
NOTRUN Referrer Policy: Expects origin for iframe-tag to cross-http origin and swap-origin redirection from http context.
NOTRUN Referrer Policy: Expects origin for iframe-tag to cross-https origin and keep-origin redirection from http context.
NOTRUN Referrer Policy: Expects origin for iframe-tag to cross-https origin and no-redirect redirection from http context.
NOTRUN Referrer Policy: Expects origin for iframe-tag to cross-https origin and swap-origin redirection from http context.
NOTRUN Referrer Policy: Expects origin for iframe-tag to same-http origin and swap-origin redirection from http context.
NOTRUN Referrer Policy: Expects origin for iframe-tag to same-https origin and keep-origin redirection from http context.
NOTRUN Referrer Policy: Expects origin for iframe-tag to same-https origin and no-redirect redirection from http context.
NOTRUN Referrer Policy: Expects origin for iframe-tag to same-https origin and swap-origin redirection from http context.
NOTRUN Referrer Policy: Expects stripped-referrer for iframe-tag to same-http origin and keep-origin redirection from http context.
NOTRUN Referrer Policy: Expects stripped-referrer for iframe-tag to same-http origin and no-redirect redirection from http context.
